updating to v1-core@0.3.13, reducing size cost of Balancer long strategy

+++ contracts/test/strategies/balancer/TestBalancerBaseStrategy.sol
@@ -49,7 +49,7 @@ contract TestBalancerBaseStrategy is BalancerBaseStrategy {
         return getSwapFeePercentage(cfmm);
     }
 
-    function testGetPoolReserves(address) public view returns(uint128[] memory) {
+    function testGetPoolReserves(address) public view returns(uint256[] memory) {
         return getPoolReserves();
     }
